Add relative duration calculation and search button
A new "Find similar melodies" button was added to **Details.cshtml**, which triggers a search using the current melody's MIDI file URL. The button's functionality is explained in an updated comment.

In **Duration.cs**, the `RelDuration()` method was introduced to calculate relative durations in quarter notes.

In **Melody.cs**, the `RelDurationsList` property was added to generate a list of relative durations for all notes in a melody.

In **Note.cs**, the `RelDur` property was added to compute relative durations using `RelDuration()`. The `DurName` property was updated to return a tuple of the note's duration symbol and name.